d_a_npc_uri almost matching (#2400)
* d_a_npc_uri almost matching

* Fix various issues

* Fix

+++ include/d/actor/d_a_npc.h
@@ -409,7 +409,7 @@ public:
 
     bool chkReverse() { return mDirection == 1; }
 
-    u16 getNumPnts() {
+    int getNumPnts() {
         dPath* path = mpRoomPath;
         return path->m_num;
     }
